qmllint: fix required property detection with alias
Iterate over all aliases that directly or indirectly points to the
required property, and check whether it is in the root file scope. If it
is in the root scope, than the required property can be satisfied by
setting the alias, and we shouldn't warn about unsatisfied required
properties int the QML component itself. Instead, we warn at the
place where the QML component is instantiated if the alias is not set.

Task-number: QTBUG-141086
Change-Id: I9237d0afd2e427558ebc2b5cbc97beb5248ec6a3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit b3f67b02b12331ad442e0a4edfd0bd4a31b03672)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>