qmlls: qInfo() when receiving EOF
Add a test to avoid QTBUG-142025 from regressing, and log a message with
qInfo() when receiving EOF to simplify debugging qmlls in the future.
This should help to tell if qmlls stopped on start because of
QTBUG-142025 or other reasons (invalid signature, or the qmlls client
trying to spawn qmlls from a filepath that does not exist for example).

The actual fix of QTBUG-142025 is in qtbase and tracked via QTBUG-142041.

Changes compared to the dev version:
* removed the cleanUp method that doesn't exist in dev anymore.
  The only other test function in this file (warnings()) already does
  everything that cleanUp attempted to do.
* make sure to disconnect also stderr from m_server, otherwise it
  segfaults in dontShutdownOnStartup() trying to reference local variable
  from warnings().

Fixes: QTBUG-142025
Change-Id: I9cd215ac02e23a1d833755eda058d72622de7318
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 0784e56cea85137cf34b5065f3877fb7777099d4)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>