QQmlListProperty: restore NRVO in toList()
Found by GCC -Wnrvo, which we intend to add to headersclean.

To fix, factor the tail part of the function into a private
helper. This preserves git history on the complete original code,
compared to the "usual" solution of wrapping the tail part of such
functions in an IILE.

Amends fa259ed4ff84a5952cdb6d2c6215e92d65afa56a.

Note: The amended commit manually cherry-picked to 6.5, but only
partially, and QQmlListProperty::toList() was not part of the pick.

While this adds a new member function to a public class, this is
forwards BC, because the added entity is a member function template.
Member function templates are never part of the ABI, even if their
class was inherited by a wholesale-exported class.

Pick-to: 6.8
Task-number: QTBUG-142146
Change-Id: I3d559c6432b636608f94f29f03203111b6a95990
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit e26ba6812cde79579d721ca7f05056a63abd530f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>